Extract pad file operations from controller
